Change tag and alias list split to use regex
All cases I've seen of tag lists have been comma separated with spaces
following each comma. In case there is an instance where the list is
simply comma-separated without the spaces, the ", " in split has been
altered to /,\s*/. The speed difference is on the order of
ten-thousandths of a second slower on lists up to 5000 tags.